Data.Table

Description

This module provides tables with multiple indices that support a simple API based on the lenses and traversals from the lens package.

Tables

data Table t where

Every Table has a Primary key and may have Candidate or Supplemental keys.

Constructors

EmptyTable :: Table t

Table :: Tabular t => Tab t (Index t) -> Table t

class Ord (PKT t) => Tabular t where

This class describes how to index a user-defined data type.

Associated Types

type PKT t

The primary key type

data Tab t m

Used to store indices

data Key k t :: * -> *

The type used internally for columns

Methods

index :: Key k t a -> t -> a

Extract the value of a Key

primary :: Key Primary t (PKT t)

Every Table has one Primary Key

primarily :: Key Primary t a -> (a ~ PKT t => r) -> r

... and so if you find one, it had better be that one!

mkTab :: Applicative h => (forall k a. (IsKeyType k, Ord a) => Key k t a -> h (i k a)) -> h (Tab t i)

Construct a Tab given a function from key to index.

ixTab :: Tab t i -> Key k t a -> i k a

Lookup an index in a Tab

forTab :: Applicative h => Tab t i -> (forall k a. (IsKeyType k, Ord a) => Key k t a -> i k a -> h (j k a)) -> h (Tab t j)

Loop over each index in a Tab

autoTab :: t -> Maybe (Tab t (Index t) -> t)

Adjust a record using meta-information about the table allowing for auto-increments.

Table Construction

empty :: Table t

Construct an empty relation

singleton :: Tabular t => t -> Table t

Construct a relation with a single row

table :: Tabular t => Iso' [t] (Table t)

Convert a list to and from a Table.

The real isomorphism laws hold if the original list makes no use of the auto-increment functionality of the table, has no duplicates and is sorted according to the primary key.

However,

from table . table ≡ id

always holds.

fromList :: Tabular t => [t] -> Table t

Build up a table from a list

Reading and Writing

null :: Table t -> Bool

Check to see if the relation is empty

count :: Table t -> Int

Retrieve a row count.

class With q t | q -> t where

Methods

with :: Ord a => q a -> (forall x. Ord x => x -> x -> Bool) -> a -> Lens' (Table t) (Table t)

Select a smaller, updateable subset of the rows of a table using an index or an arbitrary function.

group :: (Indexable a p, Applicative f, Ord a) => q a -> IndexedLensLike' p f (Table t) (Table t)

Group by a given key or arbitrary function.

deleteWith :: Ord a => q a -> (forall x. Ord x => x -> x -> Bool) -> a -> Table t -> Table t

Delete selected rows from a table

deleteWith p cmp a t ≡ set (with p cmp a) empty t

insert :: Tabular t => t -> Table t -> Table t

Insert a row into a relation, removing collisions.

delete :: t -> Table t -> Table t

Delete this row from the database. This will remove any row that collides with the specified row on any primary or candidate key.

rows :: Tabular t => Traversal (Table s) (Table t) s t

Traverse all of the rows in a table, potentially changing table types completely.

rows' :: Traversal' (Table t) t

Traverse all of the rows in a table without changing any types

Esoterica

data Auto a

Constructors

Auto !Int a

autoKey :: Lens' (Auto a) Int

auto :: a -> Auto a

Generate a row with an auto-incremented key

data Value a

Constructors

Value a

autoIncrement :: (Tabular t, PKT t ~ Int) => Loupe' t Int -> t -> Maybe (Tab t (Index t) -> t)

This lets you define autoKey to increment to 1 greater than the existing maximum key in a table.

In order to support this you need a numeric primary key, and the ability to update the primary key in a record, indicated by a lens to the field.

To enable auto-increment for a table with primary key primaryKeyField, set:

autoKey = autoIncrement primaryKeyField

Implementation Details

class IsKeyType k where

Methods

keyType :: Key k t a -> KeyType k

data KeyType t where

Value-level key types

Constructors

Primary :: KeyType Primary

Candidate :: KeyType Candidate

Supplemental :: KeyType Supplemental

data Primary

Type level key types

data Candidate

data Supplemental

data Index t k a where

This is used to store a single index.

Constructors

PrimaryIndex :: Map (PKT t) t -> Index t Primary a

CandidateIndex :: Ord a => Map a t -> Index t Candidate a

SupplementalIndex :: Ord a => Map a [t] -> Index t Supplemental a