Text.TSV.ByteString

Description

Parser for tab separated value files.

Follows naming conventions of bytestring-csv package, but treats adjacent separators as enclosing empty cells.

TSV files are a variant of comma separated value files, usually used to store tables of data organized into columns and rows.

Documentation

type Row = [ByteString]

A single row consists of a bunch of cells.

parseTSV :: ByteString -> [Row]

Parses the contents of a TSV file.

Treats \t as column separators and \n or \r\n as row separators. Several separators in a row are treated as surrounding empty columns/rows.

No column/row count consistency enforced.