Text.Tables

Description

Library for loading data from flat files. Inspired by the 'Records are Categories' Galois Tech Talk: http://www.galois.com/blog/2010/05/27/tech-talk-categories-are-databases/

Database tables and Haskell record types have uncanny similarities. A powerful analogy can be drawn:

table rows are like instances of a type

entire tables are like the type itself

columns are functions from a type to the column type

We use this analogy to easily generate text-to-types parsers.

Type-polymorphic parser

class Field a where

Class of types that can be parsed by a decoder.

Class is used to give decode a polymorphic return type, which allows simple tables to be parsed purely based on the order and types of the fields of the equivalent data type.

Methods

dec :: Decoder a

Utility functions

skip :: Int -> Decoder ()

Skipping columns of no interest.

Row parser monad

data Decoder a

Monad for parsing lists of ByteString into data structures. Parsers for individual cells are combined to build larger types, until a parser for an entire table is assembled.

runDecoder :: Decoder a -> [ByteString] -> a

Parse a single row of a table. Since flat files are usually loaded on startup, little effort is made to elegantly handle failure. If something's wrong with the tables, the program won't run long anyway.

Example: /etc/passwd

The /etc/passwd file has a clear table structure, and an equivalent type:

 data Passwd = Passwd
    { pwdAccount :: ByteString
    , pwdHash :: Hash
    , pwdUid, pwdGid :: Int
    , pwdDescription :: ByteString
    , pwdHome :: ByteString
    , pwdShell :: ByteString
    } deriving (Show,Eq)

 instance Field Passwd where
    dec = Passwd<$>dec<*>dec<*>dec<*>dec<*>dec<*>dec<*>dec

We also defined a parser that decodes rows in /etc/passwd, purely based on the types and order of elements in Passwd. It was easy because of the close relation between table and type:

All possible table entries are represented by the Passwd type.

Specific entries are represented by type instances.

Columns of the table are represented by the pwdFoo functions.

Note that we defined dec in terms of dec. We can compose column parsers from smaller column parsers. (There's a Field instance for tuples that does this.) We can also extract additional information through custom Field instances.

For example, the pwdHash column is just a ByteString but by convention the values ''*'' and ''x'' have special meaning. Why not make that clear?

 data Hash = ShadowFile | InactiveAcct | Hash ByteString deriving (Show,Eq)

 instance Field Hash where
    dec = dec >>= return . lup where
        lup "x" = ShadowFile -- (needs -XOverloadedStrings)
        lup "*" = InactiveAcct
        lup other = Hash other

You could take this further - pwdHome and pwdShell aren't ByteStrings, they're paths. What path-specific sanitization might you want to do in such a situation? When dealing with massive, 100-column tables, this kind of type composition is a lifesaver. Even with a simple passwd file, it quickly produces nice results:

>>> let splitPasswd = map (map pack . split ':') . split '\n'
>>> table <- fmap splitPasswd (readFile "/etc/passwd")
>>> let passwds = map (runDecoder dec) table
>>> passwds !! 0
Passwd { pwdAccount = "root", pwdHash = ShadowFile, pwdUid = 0, pwdGid = 0,
pwdDescription = "root", pwdHome = "/root", pwdShell = "/bin/bash"}

Example: 41-column TSV file

Take a tab-separated file with 41 columns, used to store game data. Many of the columns form patterns, the file is a flat representation of structured data!

Here's a type equivalent that describes all 41 columns with 11 properties:

 data Gem = Gem
     { name :: ByteString
     , letter :: Maybe ByteString
     , transform :: Maybe Int
     , code :: ByteString
     , nummods :: Int
     , weaponMods :: [Mod]
     , helmMods :: [Mod]
     , shieldMods :: [Mod]
     } deriving (Show,Eq)

 data Mod = Mod
     { modCode :: ByteString
     , modParam :: Int
     , modRange :: (Int,Int)
     } deriving (Show,Eq)

Most of the structure is defined by the type: the parser is defined via a boilerplate `instance Field Type where TypeConstr$dec*dec*dec*...` line.

Some information can't fit into the type, however. The lists, for instance, represent repeating columns - but how many? Here, using dec to parse isn't enough. And while it's clear what Maybe Foo means, what value of Foo counts as Nothing?

 instance Field Gem where
     dec = Gem<$>dec<*>dec<*>dec<*>dec<*>dec<*>mods<*>mods<*>mods where
         -- `[Mod]` is parsed by parsing 3 `Maybe Mod` values and reducing
         mods = replicateM 3 dec >>= return . catMaybes

 instance Field (Maybe Mod) where
     dec = do
         mod <- Mod <$> dec <*> dec <*> dec
         -- here, again, we add additional info not present in the type of Mod:
         return $ guard (modCode mod /= S.empty) >> Just mod